test(organizations): move httpError helper out of __tests__
The httpError test helper was the only non-.test.ts file inside a command
__tests__ folder. On Windows the swc --ignore globs aren't applied (cmd.exe
does not strip the quotes), so __tests__ files land in dist/. oclif manifest
filters out *.test.js but treated dist/commands/organizations/__tests__/
httpError.js as a command (organizations:__tests__:httpError), failing the
build and every Windows unit/integration job. Move it to test/helpers/,
matching where other shared test helpers live and outside the commands tree.
