+++ packages/@sanity/cli/src/commands/organizations/__tests__/delete.test.ts
@@ -2,8 +2,8 @@ import {input} from '@sanity/cli-core/ux'
 import {testCommand} from '@sanity/cli-test'
 import {afterEach, describe, expect, test, vi} from 'vitest'
 
+import {httpError} from '../../../../test/helpers/httpError.js'
 import {DeleteOrganizationCommand} from '../delete.js'
-import {httpError} from './httpError.js'
 
 const mockRequest = vi.hoisted(() => vi.fn())
